requirements.txt
setuptools
anthropic==0.37.1
beautifulsoup4==4.12.3
boto3==1.35.48
botocore==1.35.48
chromadb==0.4.24
cohere==5.11.1
duckduckgo_search==6.3.2
fireworks-ai==0.15.7
groq==0.11.0
huggingface-hub==0.26.1
jinja2==3.1.4
langchain==0.3.4
langchain-anthropic==0.2.3
langchain-cohere==0.3.1
langchain-community==0.3.3
langchain-core==0.3.12
langchain-experimental==0.3.2
langchain-fireworks==0.2.1
langchain-google-vertexai==2.0.5
langchain-groq==0.2.0
langchain-mistralai==0.2.0
langchain-nvidia-ai-endpoints==0.3.1
langchain-openai==0.2.3
langgraph==0.2.39
langgraph-checkpoint==2.0.2
langgraph-checkpoint-sqlite==2.0.1
neo4j
numpy
openai
pandas
pillow
playwright
playwright-stealth
py2neo
pydantic>=2.9.2
PyPDF2
python-dotenv
requests
streamlit
tqdm
watchtower
tokenizers
transformers
tenacity
sqlalchemy
psycopg2
psycopg2-binary
pybloom_live
fastapi
uvicorn
python-multipart